requirements.txt
discord.py==1.7.3
beautifulsoup4==4.10.0